+++ packages/ai-jsx/src/lib/tts/tts.tsx
@@ -308,6 +308,23 @@ export class AwsTextToSpeech extends RestTextToSpeech {
 }
 
 /**
+ * Text-to-speech implementation that uses the Google Cloud text-to-speech service.
+ */
+export class GcpTextToSpeech extends RestTextToSpeech {
+  static readonly DEFAULT_VOICE = 'en-US-Wavenet-D';
+  constructor(urlFunc: BuildUrl, voice: string = GcpTextToSpeech.DEFAULT_VOICE, rate: number = 1.0) {
+    super('gcp', urlFunc, voice, rate);
+  }
+}
+
+export class WellSaidTextToSpeech extends RestTextToSpeech {
+  static readonly DEFAULT_VOICE = '42'; // Sofia H. (Conversational)
+  constructor(urlFunc: BuildUrl, voice: string = WellSaidTextToSpeech.DEFAULT_VOICE, rate: number = 1.0) {
+    super('wellsaid', urlFunc, voice, rate);
+  }
+}
+
+/**
  * Text-to-speech implementation that uses a web socket to stream text to the
  * server and receives audio chunks as they are generated.
  */
@@ -419,6 +436,10 @@ class ElevenLabsOutboundMessage {
     this.xi_api_key = xi_api_key;
   }
   text: string;
+  voice_settings?: {
+    stability: number;
+    similarity: number;
+  };
   generation_config?: {
     chunk_length_schedule: number[];
   };
@@ -443,6 +464,10 @@ export class ElevenLabsTextToSpeech extends WebSocketTextToSpeech {
     // once we have 50 characters of text buffered.
     const obj = new ElevenLabsOutboundMessage({
       text: ' ',
+      voice_settings: {
+        stability: 0.5,
+        similarity: 0.8,
+      },
       generation_config: {
         chunk_length_schedule: [50],
       },
@@ -454,13 +479,11 @@ export class ElevenLabsTextToSpeech extends WebSocketTextToSpeech {
     const message = inMessage as ElevenLabsInboundMessage;
     console.debug(message);
     if (message.audio) {
-      if (!message.isFinal) {
-        console.debug(`[${this.name}] chunk received`);
-        this.queueChunk(new AudioChunk(Buffer.from(message.audio!, 'base64')));
-      } else {
-        console.log(`[${this.name}] utterance complete`);
-        this.setComplete();
-      }
+      console.debug(`[${this.name}] chunk received`);
+      this.queueChunk(new AudioChunk(Buffer.from(message.audio!, 'base64')));
+    } else if (message.isFinal) {
+      console.log(`[${this.name}] utterance complete`);
+      this.setComplete();
     } else if (message.error) {
       console.error(`[${this.name}] error: ${message.message}`);
     }
@@ -486,11 +509,18 @@ export class TextToSpeechOptions {
  * Factory function to create a text-to-speech service for the specified provider.
  */
 export function createTextToSpeech({ provider, rate, voice, getToken, buildUrl }: TextToSpeechOptions) {
-  if (provider == 'azure') {
-    return new AzureTextToSpeech(buildUrl!, voice, rate);
+  switch (provider) {
+    case 'azure':
+      return new AzureTextToSpeech(buildUrl!, voice, rate);
+    case 'aws':
+      return new AwsTextToSpeech(buildUrl!, voice, rate);
+    case 'gcp':
+      return new GcpTextToSpeech(buildUrl!, voice, rate);
+    case 'wellsaid':
+      return new WellSaidTextToSpeech(buildUrl!, voice, rate);
+    case 'eleven':
+      return new ElevenLabsTextToSpeech(getToken!, voice);
+    default:
+      throw new Error(`unknown provider ${provider}`);
   }
-  if (provider == 'eleven') {
-    return new ElevenLabsTextToSpeech(getToken!, voice);
-  }
-  throw new Error(`unknown provider ${provider}`);
 }